Fix in ferry. Started exercise 4. Not done with it

+++ src/jt222ii_assign2/Exercise2/FerrySys.java
@@ -67,7 +67,7 @@ public class FerrySys implements Ferry
     @Override
     public boolean hasSpaceFor(Vehicle v) {
 
-        return Math.ceil((int)occupiedSpace+v.getSpace()) <= maxSpace && !vehicles.contains(v);
+        return Math.ceil(occupiedSpace+v.getSpace()) <= maxSpace && !vehicles.contains(v);
     }
 
     @Override
